+++ build.gradle
@@ -13,41 +13,18 @@ configurations {
 		extendsFrom annotationProcessor
 	}
 }
-
 repositories {
 	mavenCentral()
 }
-ext{
-	set('springCloudVersion',"2021.0.1")
-}
 
 dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-jdbc'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation 'mysql:mysql-connector-java'
-
-	testImplementation 'com.h2database:h2'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-
-	runtimeOnly 'mysql:mysql-connector-java'
-
 	compileOnly 'org.projectlombok:lombok'
+	runtimeOnly 'com.h2database:h2'
+	runtimeOnly 'mysql:mysql-connector-java'
 	annotationProcessor 'org.projectlombok:lombok'
-
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-
-	implementation 'io.springfox:springfox-boot-starter:3.0.0'
-	implementation 'org.springframework.data:spring-data-envers'
-	implementation 'org.apache.commons:commons-lang3:3.12.0'
-
-	implementation project(path: ":domain",configuration: 'default')
-}
-
-dependencyManagement {
-	imports{
-		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-	}
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
 tasks.named('test') {
