+++ src/main/java/com/zb/cms/CmsApplication.java
@@ -1,19 +1,9 @@
 package com.zb.cms;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.servlet.ServletComponentScan;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@EnableFeignClients
 @SpringBootApplication
-@RequiredArgsConstructor
-@EnableJpaAuditing
-@EnableJpaRepositories
-@ServletComponentScan
 public class CmsApplication {
 
 	public static void main(String[] args) {
