--- src/main/java/com/zb/cms/client/MailgunClient.java
@@ -1,18 +1,0 @@
-package com.zb.cms.client;
-
-import com.zb.cms.client.mailgun.SendMailForm;
-import feign.Response;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.cloud.openfeign.SpringQueryMap;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-
-@FeignClient(name = "mailgun", url = "https://api.mailgun.net/v3/")
-@Qualifier("mailgun")
-public interface MailgunClient {
-
-    @PostMapping("sandbox2d2a0219b48547898ba0619e4480a8d9.mailgun.org/messages")
-    ResponseEntity<String> sendEmail(@SpringQueryMap SendMailForm form);
-
-}
