--- src/main/java/com/zb/cms/config/JwtConfig.java
@@ -1,13 +1,0 @@
-package com.zb.cms.config;
-
-import com.zb.domain.config.JwtAuthenticationProvider;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JwtConfig {
-    @Bean
-    public JwtAuthenticationProvider jwtAuthenticationProvider(){
-        return new JwtAuthenticationProvider();
-    }
-}
