--- src/main/java/com/zb/cms/controller/SignInController.java
@@ -1,23 +1,0 @@
-package com.zb.cms.controller;
-
-import com.zb.cms.application.SignInApplication;
-import com.zb.cms.domain.repository.SignInForm;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/signIn")
-@RequiredArgsConstructor
-public class SignInController {
-
-    private final SignInApplication signInApplication;
-
-    @PostMapping("/customer")
-    public ResponseEntity<String> signInCustomer(@RequestBody SignInForm form){
-        return ResponseEntity.ok(signInApplication.customerLoginToken(form));
-    }
-}
