--- src/main/java/com/zb/cms/controller/TestController.java
@@ -1,18 +1,0 @@
-package com.zb.cms.controller;
-
-import com.zb.cms.service.test.EmailSendService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class TestController {
-    private final EmailSendService emailSendService;
-
-    @GetMapping
-    public String sendTestEmail(){
-        return emailSendService.sendEmail();
-    }
-
-}
