--- src/main/java/com/zb/cms/domain/model/BaseEntity.java
@@ -1,21 +1,0 @@
-package com.zb.cms.domain.model;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(value = {AuditingEntityListener.class})
-public abstract class BaseEntity {
-
-    @CreatedDate
-    private LocalDateTime createAt;
-    @LastModifiedDate
-    private LocalDateTime modifiedAt;
-}
