--- src/main/java/com/zb/cms/domain/repository/CustomerRepository.java
@@ -1,13 +1,0 @@
-package com.zb.cms.domain.repository;
-
-import com.zb.cms.domain.model.Customer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface CustomerRepository extends JpaRepository<Customer,Long> {
-    Optional<Customer> findByEmail(String email);
-
-}
