--- src/main/java/com/zb/cms/domain/repository/SignInForm.java
@@ -1,10 +1,0 @@
-package com.zb.cms.domain.repository;
-
-import lombok.Getter;
-
-@Getter
-
-public class SignInForm {
-    private String email;
-    private String password;
-}
