--- src/main/java/com/zb/cms/exception/CustomException.java
@@ -1,14 +1,0 @@
-package com.zb.cms.exception;
-
-import com.zb.cms.domain.model.Customer;
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getDetail());
-        this.errorCode = errorCode;
-    }
-}
