--- src/main/java/com/zb/cms/service/test/EmailSendService.java
@@ -1,25 +1,0 @@
-package com.zb.cms.service.test;
-
-import com.zb.cms.client.MailgunClient;
-import com.zb.cms.client.mailgun.SendMailForm;
-import feign.Response;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class EmailSendService {
-    private final MailgunClient mailgunClient;
-
-    public String sendEmail(){
-
-        SendMailForm form = SendMailForm.builder()
-                .from("onlee1425@naver.com")
-                .to("onlee1425@naver.com")
-                .subject("Test email from zb cms")
-                .text("Test text")
-                .build();
-
-        return mailgunClient.sendEmail(form).getBody();
-    }
-}
