--- src/test/java/com/zb/cms/client/service/EmailSendServiceTest.java
@@ -1,22 +1,0 @@
-package com.zb.cms.client.service;
-
-import com.zb.cms.client.MailgunClient;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest()
-class EmailSendServiceTest {
-
-    @Autowired
-    private MailgunClient mailgunClient;
-
-    @Test
-    public void EmailTest(){
-        // need test code
-        mailgunClient.sendEmail(null);
-//        String response = emailSendService.sendEmail();
-//        System.out.println(response);
-    }
-
-}
