--- src/test/java/com/zb/cms/service/SignUpCustomerServiceTest.java
@@ -1,33 +1,0 @@
-package com.zb.cms.service;
-
-import com.zb.cms.domain.SignUpForm;
-import com.zb.cms.domain.model.Customer;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.util.Assert;
-
-import java.time.LocalDate;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-@SpringBootTest
-class SignUpCustomerServiceTest {
-
-    @Autowired
-    private SignUpCustomerService service;
-
-    @Test
-    void signUp(){
-        SignUpForm form = SignUpForm.builder()
-                .name("name")
-                .birth(LocalDate.now())
-                .email("abc@gmail.com")
-                .password("1")
-                .phone("010-0000-0000")
-                .build();
-        Customer c = service.signUp(form);
-       assertNotNull(c.getId());
-       assertNotNull(c.getCreateAt());
-    }
-}
