+++ user-api/src/main/java/com/zb/cms/user/UserApplication.java
@@ -1,0 +1,23 @@
+package com.zb.cms.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.ServletComponentScan;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@EnableFeignClients
+@SpringBootApplication
+@RequiredArgsConstructor
+@EnableJpaAuditing
+@EnableJpaRepositories
+@ServletComponentScan
+public class UserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserApplication.class, args);
+    }
+
+}
