+++ user-api/src/main/java/com/zb/cms/user/client/mailgun/SendMailForm.java
@@ -1,0 +1,17 @@
+package com.zb.cms.user.client.mailgun;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Getter
+public class SendMailForm {
+    private String from;
+    private String to;
+    private String subject;
+    private String text;
+}
