+++ user-api/src/main/java/com/zb/cms/user/config/FeignConfig.java
@@ -1,0 +1,20 @@
+package com.zb.cms.user.config;
+
+import feign.auth.BasicAuthRequestInterceptor;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class FeignConfig {
+
+    @Value(value = "${mailgun.key}")
+    private String mailgunKey;
+
+    @Qualifier(value = "mailgun") //mailgun 에서만 사용
+    @Bean
+    public BasicAuthRequestInterceptor basicAuthRequestInterceptor(){
+        return new BasicAuthRequestInterceptor("api",mailgunKey);
+    }
+}
