+++ user-api/src/main/java/com/zb/cms/user/config/JwtConfig.java
@@ -1,0 +1,13 @@
+package com.zb.cms.user.config;
+
+import com.zb.domain.config.JwtAuthenticationProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+    @Bean
+    public JwtAuthenticationProvider jwtAuthenticationProvider(){
+        return new JwtAuthenticationProvider();
+    }
+}
