+++ user-api/src/main/java/com/zb/cms/user/config/filter/CustomerFilter.java
@@ -1,0 +1,34 @@
+package com.zb.cms.user.config.filter;
+
+import com.zb.cms.user.service.CustomerService;
+import com.zb.domain.config.JwtAuthenticationProvider;
+import com.zb.domain.domain.common.UserVo;
+import lombok.RequiredArgsConstructor;
+
+import javax.servlet.*;
+import javax.servlet.annotation.WebFilter;
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+@WebFilter(urlPatterns = "/customer/+")
+@RequiredArgsConstructor
+public class CustomerFilter implements Filter {
+
+    private final JwtAuthenticationProvider jwtAuthenticationProvider;
+    private final CustomerService customerService;
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
+            throws IOException, ServletException {
+        HttpServletRequest req = (HttpServletRequest) request;
+        String token = req.getHeader("X-AUTH-TOKEN");
+        if (!jwtAuthenticationProvider.validateToken(token)) {
+            throw new ServletException("Invalid Access");
+        }
+        UserVo vo = jwtAuthenticationProvider.getUserVo(token);
+        customerService.findByIdAndEmail(vo.getId(), vo.getEmail()).orElseThrow(
+                () -> new ServletException("Invalid access")
+        );
+        chain.doFilter(request, response);
+    }
+}
