+++ user-api/src/main/java/com/zb/cms/user/controller/TestController.java
@@ -1,0 +1,18 @@
+package com.zb.cms.user.controller;
+
+import com.zb.cms.user.service.test.EmailSendService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class TestController {
+    private final EmailSendService emailSendService;
+
+    @GetMapping
+    public String sendTestEmail(){
+        return emailSendService.sendEmail();
+    }
+
+}
