+++ user-api/src/main/java/com/zb/cms/user/domain/SignUpForm.java
@@ -1,0 +1,20 @@
+package com.zb.cms.user.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignUpForm {
+    private String email;
+    private String name;
+    private String password;
+    private LocalDate birth;
+    private String phone;
+}
