+++ user-api/src/main/java/com/zb/cms/user/domain/model/Customer.java
@@ -1,0 +1,49 @@
+package com.zb.cms.user.domain.model;
+
+import com.zb.cms.user.domain.SignUpForm;
+import lombok.*;
+import org.hibernate.envers.AuditOverride;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.util.Locale;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@AuditOverride(forClass = BaseEntity.class)
+public class Customer extends BaseEntity {
+    @Id
+    @Column(name = "id", nullable = false)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String email;
+
+    private String name;
+    private String password;
+    private String phone;
+    private LocalDate birth;
+
+    //email verify
+    private LocalDateTime verifyExpiredAt;
+    private String verificationCode;
+    private boolean verify;
+
+    public static Customer from(SignUpForm form) {
+        return Customer.builder()
+                .email(form.getEmail().toLowerCase(Locale.ROOT))
+                .password(form.getPassword())
+                .name(form.getName())
+                .birth(form.getBirth())
+                .phone(form.getPhone())
+                .verify(false)
+                .build();
+    }
+
+}
