+++ user-api/src/main/java/com/zb/cms/user/domain/repository/CustomerRepository.java
@@ -1,0 +1,13 @@
+package com.zb.cms.user.domain.repository;
+
+import com.zb.cms.user.domain.model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer,Long> {
+    Optional<Customer> findByEmail(String email);
+
+}
