+++ user-api/src/main/java/com/zb/cms/user/domain/repository/SignInForm.java
@@ -1,0 +1,10 @@
+package com.zb.cms.user.domain.repository;
+
+import lombok.Getter;
+
+@Getter
+
+public class SignInForm {
+    private String email;
+    private String password;
+}
