+++ user-api/src/main/java/com/zb/cms/user/exception/CustomException.java
@@ -1,0 +1,13 @@
+package com.zb.cms.user.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public CustomException(ErrorCode errorCode) {
+        super(errorCode.getDetail());
+        this.errorCode = errorCode;
+    }
+}
