+++ user-api/src/main/java/com/zb/cms/user/exception/ErrorCode.java
@@ -1,0 +1,23 @@
+package com.zb.cms.user.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+@Getter
+public enum ErrorCode {
+    //REGISTER
+    ALREADY_REGISTER_USER(HttpStatus.BAD_REQUEST,"이미 가입된 회원입니다."),
+    NOT_FOUND_USER(HttpStatus.BAD_REQUEST,"일치하는 회원이 없습니다."),
+    ALREADY_VERIFY(HttpStatus.BAD_REQUEST,"이미 인증이 완료되었습니다"),
+    WRONG_VERIFICATION(HttpStatus.BAD_REQUEST,"잘못된 인증 시도입니다."),
+    EXPIRE_CODE(HttpStatus.BAD_REQUEST,"인증 시간이 만료되었습니다."),
+
+    //LOGIN
+    LOGIN_CHECK_FAIL(HttpStatus.BAD_REQUEST,"아이디 또는 패스워드를 확인 해 주세요.");
+
+    private final HttpStatus httpStatus;
+    private final String detail;
+
+}
