+++ user-api/src/main/resources/application.properties
@@ -1,0 +1,11 @@
+spring.mvc.pathmatch.matching-strategy=ant_path_matcher
+feign.okhttp.enabled=true
+
+#spring.datasource.url=jdbc:mysql://localhost:3306/cms?serverTimezone=UTC&characterEncoding=UTF-8
+spring.datasource.url=jdbc:mysql://localhost:3306/cms?useSSL=false&useUnicode=true&allowPublicKeyRetrieval=true
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.username=root
+spring.datasource.password=1
+
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=create
